@@ -1,16 +1,17 @@
 require('../styles/main.css');
-const pokedex = require('json!./pokedex.json');
+import { pokedexDE, pokedexEN } from 'pokedex.js';
 let pokemonLeft = [];
 const image = document.querySelector('.shape');
 const guess = document.querySelector('#guess');
 const score = document.querySelector('.score');
+let pokedex = pokedexDE;
 
 function renderImage(pokemonId) {
   image.src = `../assets/${pokemonId}.png`;
 }
 
 function getPokemon() {
-  if(pokemonLeft) {
+  if (pokemonLeft) {
     const pokemonNr = Math.floor(Math.random() * (pokemonLeft.length - 1));
     const pokemon = pokemonLeft[pokemonNr];
     pokemonLeft = [...pokemonLeft.slice(0, pokemonNr),
@@ -22,17 +23,14 @@ function getPokemon() {
 
 function initGame() {
   score.innerHTML = 0;
-  for(let it = 1; it <= 151; it++) {
+  for (let it = 1; it <= 151; it++) {
     pokemonLeft.push(it);
   }
   renderImage(getPokemon());
 }
-
-initGame();
-
 guess.addEventListener('keypress', (event) => {
-  if(event.which === 13) {
-    if(pokedex[image.src.split('/').pop().split('.').shift()] === guess.value.toUpperCase()) {
+  if (event.which === 13) {
+    if (pokedex[image.src.split('/').pop().split('.').shift()] === guess.value.toUpperCase()) {
       image.classList.toggle('darken');
       score.innerHTML = parseInt(score.innerHTML, 10) + 1;
       setTimeout(() => {
@@ -43,3 +41,5 @@ guess.addEventListener('keypress', (event) => {
     }
   }
 });
+
+initGame();
